Move WeaponTrainingStatsRequest class into WeaponStatsRequest.cs
